pyproject.toml
[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[project]
name = "metalake"
version = "0.0.0"
description = "Metalake to store any data"
authors = [
    { name = "Roman", email = "iamromandev@gmail.com" },
]
readme = "README.md"
requires-python = "==3.13.5"
dependencies = [
    "loguru==0.7.3",
    "fastapi[all]==0.117.1",
    "tortoise-orm[accel,asyncmy]==0.25.1",
    "aerich[toml,asyncmy]==0.9.1",
    "toml==0.10.2",
    "minio==7.2.17",
]

[dependency-groups]
dev = [
    "pre-commit==4.3.0",
    "ruff==0.13.2",
    "mypy==1.18.2",
    "ty==0.0.1a21",
]

[tool.uv]
package = false

[tool.aerich]
tortoise_orm = "src.db.DB_CONFIG"
location = "./src/db/migrations"
src_folder = "./."
